ui: improvements proposed by Julian Mautner
* automatically expand active class
* increase arrow size
* return child nodes in autocompletions
* always return filters in autocompletions

+++ src/app/qt/view/graphElements/QtGraphEdge.h
@@ -77,7 +77,7 @@ protected:
 private:
 	bool isAggregation() const;
 	int getZValue(bool active) const;
-	int getPenWidth() const;
+	float getPenWidth() const;
 	int getAggregationCount() const;
 
 	std::weak_ptr<GraphNode> m_owner;
